Move Task files in seperate folders
