+++ imports/client/ui/tasks/Memory/MemoryAnimationTemplate.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Meteor } from "meteor/meteor";
 import PropTypes from "prop-types";
-import { DragdropModel } from "../../../models/DragdropModel";
+import { DragdropModel } from "../../../../models/DragdropModel";
 import { Button, Header, Modal, Image, Icon, List } from "semantic-ui-react";
 
 import Swal from "sweetalert2";
