+++ imports/client/ui/tasks/MultiChoice/MultiChoiceView.js
